[PURPLE-135] EvaluationField, EvaluationOption Enum 전환으로 인한 리팩토링
