+++ src/main/java/com/pikachu/purple/application/evaluation/port/in/EvaluationFieldAndEvaluationMoodGetUseCase.java
@@ -4,7 +4,7 @@ import com.pikachu.purple.application.evaluation.common.dto.EvaluationFieldDTO;
 import com.pikachu.purple.domain.evaluation.EvaluationMood;
 import java.util.List;
 
-public interface EvaluationCodeGetUseCase {
+public interface EvaluationFieldAndEvaluationMoodGetUseCase {
 
     Result invoke();
