--- src/main/java/com/pikachu/purple/application/evaluation/service/domain/EvaluationCodeDomainService.java
@@ -1,10 +1,0 @@
-package com.pikachu.purple.application.evaluation.service.domain;
-
-import com.pikachu.purple.domain.evaluation.EvaluationCode;
-import java.util.List;
-
-public interface EvaluationCodeDomainService {
-
-    List<EvaluationCode> findAll();
-
-}
