--- src/main/java/com/pikachu/purple/application/evaluation/service/domain/EvaluationFieldOptionDomainService.java
@@ -1,10 +1,0 @@
-package com.pikachu.purple.application.evaluation.service.domain;
-
-import com.pikachu.purple.domain.evaluation.EvaluationFieldOption;
-import java.util.List;
-
-public interface EvaluationFieldOptionDomainService {
-
-    List<EvaluationFieldOption> findAll();
-
-}
