+++ src/main/java/com/pikachu/purple/bootstrap/review/controller/ReviewController.java
@@ -1,7 +1,7 @@
 package com.pikachu.purple.bootstrap.review.controller;
 
-import com.pikachu.purple.application.evaluation.port.in.EvaluationCodeGetUseCase;
-import com.pikachu.purple.application.evaluation.port.in.EvaluationCodeGetUseCase.Result;
+import com.pikachu.purple.application.evaluation.port.in.EvaluationFieldAndEvaluationMoodGetUseCase;
+import com.pikachu.purple.application.evaluation.port.in.EvaluationFieldAndEvaluationMoodGetUseCase.Result;
 import com.pikachu.purple.application.review.port.in.ReviewCreateSimpleUseCase;
 import com.pikachu.purple.application.review.port.in.ReviewDeleteUseCase;
 import com.pikachu.purple.application.review.port.in.ReviewUpdateUseCase;
@@ -9,7 +9,7 @@ import com.pikachu.purple.bootstrap.common.dto.SuccessResponse;
 import com.pikachu.purple.bootstrap.review.api.ReviewApi;
 import com.pikachu.purple.bootstrap.review.dto.request.CreateReviewSimpleRequest;
 import com.pikachu.purple.bootstrap.review.dto.request.UpdateReviewSimpleRequest;
-import com.pikachu.purple.bootstrap.review.dto.response.GetReviewDetailInfoResponse;
+import com.pikachu.purple.bootstrap.review.dto.response.GetEvaluationFieldAndEvaluationMoodResponse;
 import lombok.RequiredArgsConstructor;
 import org.springframework.web.bind.annotation.RestController;
 
@@ -18,7 +18,7 @@ import org.springframework.web.bind.annotation.RestController;
 public class ReviewController implements ReviewApi {
 
     private final ReviewCreateSimpleUseCase reviewCreateSimpleUseCase;
-    private final EvaluationCodeGetUseCase evaluationCodeGetUseCase;
+    private final EvaluationFieldAndEvaluationMoodGetUseCase evaluationFieldAndEvaluationMoodGetUseCase;
     private final ReviewUpdateUseCase reviewUpdateUseCase;
     private final ReviewDeleteUseCase reviewDeleteUseCase;
 
@@ -34,10 +34,10 @@ public class ReviewController implements ReviewApi {
     }
 
     @Override
-    public SuccessResponse<GetReviewDetailInfoResponse> getDetailInfo() {
-        Result result = evaluationCodeGetUseCase.invoke();
+    public SuccessResponse<GetEvaluationFieldAndEvaluationMoodResponse> findEvaluationFieldAndEvaluationMood() {
+        Result result = evaluationFieldAndEvaluationMoodGetUseCase.invoke();
 
-        return SuccessResponse.of(new GetReviewDetailInfoResponse(
+        return SuccessResponse.of(new GetEvaluationFieldAndEvaluationMoodResponse(
             result.evaluationFieldDTOS(),
             result.evaluationMoods()
         ));
